ReferenceAppAgent
This agent controls a set of simulated device drivers, and
communicates with an OpenADRVenAgent that relays demand-response
event information. During a DR event, it reduces simulated power
demand. Devices are controlled via an ActuatorAgent.

Includes a set of regression tests.

Documentation for this app has been added to VOLTTRON ReadTheDocs
under applications.